requirements.txt
fastapi
uvicorn[standard]
pydantic
torch
transformers
huggingface_hub[cli]
bitsandbytes
accelerate
pyyaml
pandas
scikit-learn
pytest
pyngrok